pyproject.toml
[project]
name = "black-marble"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = ["httpx", "python-dotenv>=1.0.1"]

[tool.uv.sources]
blackmarblepy = { path = "../blackmarblepy", editable = true }
